Arca_musarithmica.Syntagma1.Pinax03a

Description

Sub-structures used by Arca_musarithmica to build Kircher's ark. This pinax (s1p3a) is for poetry in Adonic meter (`--`-) with five syllables per line. It contains four columns, one for each successive verse line ("Kircher's stropha") of a quatrain. There is only one rperm for each side, repeated for each column.

It may be used with any tone, especially 2 and 6.

Documentation

s1p3a :: Pinax

Pinax 3