Cogito.Musarithmetic

Description

This module provides the tools used in the main Cogito module to adjust music created by the ark and to store it in internal structures that will then be used by the Scribo modules.

Pitches and Groups of Them

data RawPitch

A RawPitch is the same as a Pitch except instead of a Pnum it has an Int. This allows us to add and subtract from pitches and then use stdPitch to normalize them by adjusting the octave and pitch like the two digits of a base-7 number.

Constructors

RawPitch

Fields

rawPnum :: Int

rawOct :: Int

rawDur :: Dur

rawAccid :: Accid

rawAccidType :: AccidType

data Voice

A Voice is a list of pitches with an identifier for the voice type.

Constructors

Voice

Fields

voiceID :: VoiceName

Enum for Soprano, Alto, Tenor or Bass

music :: [Pitch]

mergeVoices :: [Voice] -> Voice

Make a list of voices with same voiceID into one Voice (combine the lists of Pitches into one)

type Chorus = [Voice]

A Chorus is a group (list) of four Voice items

TODO: We don't actually define it as being four items. TODO: Do we still need this with new MEI setup?

Manipulating the Pitch

newRest

Arguments

:: Dur
Rhythmic duration for this note

-> Pitch

Create a rest (that is, a Pitch with duration only)

We make a Pitch but set the pnum to Rest; oct and accid are set to special nil values (OctNil, AccidNil)

TODO: We are setting the octave using fromEnum OctNil: Isn't this the same as just setting it to zero? Is there a better way to mark this?

stdPitch :: RawPitch -> Pitch

Standardize pitch.

A Pitch is like a two-digit number in base 7, where the first digit (the "7s") is the octave. If the second digit (the "1s") is over 7, we must increment the octave. A "standardized pitch" in our implementation is one that conforms to this requirement, so the stdPitch method takes all the input for a Pitch and does the necessary conversions to set the octave.

We need this because Kircher's tables include both pitch number 1 and pitch 8. If pitch input is Kircher's pitch 8, then set pitch num to PcC (0) and add one to octave.

TODO: We don't check for values out of range because we know what input we are getting from the ark. Is this okay?

incPitch :: Pitch -> Int -> Pitch

Increment a pitch by increasing its Pnum pitch number and its octave if necessary (using Pitch structure like a base-7 number). Create a new Pitch with incremented Pnum and then standardize it with stdPitch to get correct octave and pitch number.

Adjust pitch for mode

modeMollis :: Mode -> ModeSystem -> Bool

Is a mode in cantus mollis? Should there be a flat in the key signature?

pnumAccidInMode :: Int -> ModeList -> Mode -> PnumAccid

Adjust a pitch to be in a given mode.

modalFinal :: ModeList -> Mode -> Pitch

Get the modal final for this mode. What pitch = 0 in this mode? (In Kircher's 1-indexed vperms, the final is 1 or 8.)

modalFinalInRange :: ModeList -> Mode -> VoiceName -> VoiceRanges -> Pitch

Get the modal final within range for this voice. What is the lowest octave of that pitch permissible in the range for that voice? Set all the other pitches with reference to that, within the octave between 0--7

modalOctaveBase :: ModeList -> Mode -> VoiceName -> VoiceRanges -> Int

What octave is the modal final in for this voice's range?

isRest :: Dur -> Bool

Check to see if a rhythmic duration is a rest type (the rest enums begin with LgR so we compare with that)

isPitchRest :: Pitch -> Bool

Is the Pitch a rest?

Measure distances between notes and correct bad intervals

Convert between diatonic and chromatic pitches to calculate intervals

absPitch :: Pitch -> Int

Convert Pitch to absolute pitch number

dia2chrom :: Pnum -> Int

Get chromatic offset from C for diatonic pitch classes

pEq :: Pitch -> Pitch -> Bool

Are two Pitches the same chromatic pitch, enharmonically equivalent?

pGt :: Pitch -> Pitch -> Bool

Pitch greater than?

pLt :: Pitch -> Pitch -> Bool

Pitch less than?

pGtEq :: Pitch -> Pitch -> Bool

Pitch greater than or equal?

pLtEq :: Pitch -> Pitch -> Bool

Pitch less than or equal?

p12diff :: Pitch -> Pitch -> Int

Difference between pitches, chromatic interval

p12diffMod :: Pitch -> Pitch -> Int

p12diff modulo 12 (= chromatic difference within one octave)

absPitch7 :: Pitch -> Int

Absolute diatonic pitch (octave + pitch)

p7diff :: Pitch -> Pitch -> Int

Difference between pitches, diatonic interval Unison = 0, therefore results of this function are one less than the verbal names of intervals (p7diff = 4 means a fifth)

p7diffMod :: Pitch -> Pitch -> Int

Diatonic difference between pitch classes (= pitch difference as though within a single octave); result is 0-indexed, so the interval of a "third" in speech has a p7diffMod of 2

octaveAdjust :: Pitch -> (Int -> Int) -> Pitch

Copy a Pitch (unchanged if Rest), with given function applied to the octave member

octaveUp :: Pitch -> Pitch

Raise the octave by 1

octaveDown :: Pitch -> Pitch

Lower the octave by 1

pitchTooLow :: Pitch -> VoiceName -> VoiceRanges -> Bool

Is the pitch below the bottom limit of the voice range?

pitchTooHigh :: Pitch -> VoiceName -> VoiceRanges -> Bool

Is the pitch above the upper limit of the voice range?

lowRange :: VoiceName -> VoiceRanges -> Pitch

Get the bottom limit of the voice range

highRange :: VoiceName -> VoiceRanges -> Pitch

Get the top limit of the voice range

adjustPitchInRange :: Pitch -> VoiceName -> VoiceRanges -> Pitch

Adjust a pitch to be in the correct voice range (using Aedifico.vocalRanges). If it's in the right range for the voice, leave it alone; if it's too low raise it by an octave, or vice versa if it's too high; keep shifting octaves till it's in range.

isPitchInRange :: Pitch -> VoiceName -> VoiceRanges -> Bool

Is the Pitch within the proper range for its voice?

Adjust list of pitches to avoid bad intervals

stepwise :: [Pitch] -> [Pitch]

Go through list of pitches and reduce intervals that are too large

stepwiseVoice :: Voice -> Voice

Adjust a whole Voice stepwise

_maxLeap :: Int

No leaps bigger than this interval This is based on 0 as unison so a _maxLeap of 4 is a fifth

unleap :: Pitch -> Pitch -> Pitch

Reduce leap of more than a _maxLeap by shifting octave of second note up or down until the interval is within range

Octave leaps are okay, though.

TODO : - but what if after adjusting for leaps, a note is out of range? - and what if there is a descending scale that goes out of range and the only way to adjust it is to make a seventh? need to adjust a whole phrase - But this may be going beyond what Kircher specifies as fully-automated algorithms.

pitchMax :: [Pitch] -> Maybe Pitch

Return the highest pitch in a list of pitches.

pitchMin :: [Pitch] -> Maybe Pitch

Return the lowest pitch in a list of pitches.

stepwiseInRange :: [Pitch] -> VoiceName -> VoiceRanges -> [Pitch]

TODO write your own max/min functions for pitches that ignore Rests

Go through list of pitches and reduce intervals that are too large

voiceInRange :: VoiceRanges -> Voice -> Voice

Adjust a whole Voice to be in range: check the highest and lowest notes in the list, compare to the range for the voice, and shift the whole thing by octave until all are in range; return error if it can't be done

stepwiseVoiceInRange :: VoiceRanges -> Voice -> Voice

Adjust a whole Voice stepwise

Data structures to store music composed by the ark

data Note

A Note contains a pitch and a syllable, equivalent to MEI note

Constructors

Note

Fields

notePitch :: Pitch

noteSyllable :: Syllable

data Syllable

A Syllable is a single syllable to be paired with a Pitch, including its position in the word.

Constructors

Syllable

Fields

sylText :: String

sylPosition :: SyllablePosition

data SyllablePosition

What is the position of the syllable relative to the word? Beginning, middle, or end? This determines hyphenation.

Constructors

First

Middle

Last

Only

Tacet
no syllable

data MusicPhrase

A MusicPhrase contains all the notes set using one permutation drawn from the ark, for a single voice.

Constructors

MusicPhrase

Fields

phraseVoiceID :: VoiceName

notes :: [Note]

type MusicSentence = [MusicPhrase]

A list of MusicPhrase items

data MusicSection

A MusicSection contains all the music for one section in the input XML document, for a single voice, together with the parameters set in the input file.

Constructors

MusicSection

Fields

secVoiceID :: VoiceName

secConfig :: ArkConfig

secSentences :: [MusicSentence]

data MusicChorus

A MusicChorus is a four-voice SATB structure of MusicSection data. TODO do we really need it to be structured this way?

Constructors

MusicChorus

Fields

soprano :: MusicSection

alto :: MusicSection

tenor :: MusicSection

bass :: MusicSection

type MusicScore = [MusicChorus]

The full MusicScore is a list of SATB MusicChorus structures.