Scribo.Lilypond

Description

This module is our implementation of Kircher's palimpsest phonotacticum, his system for writing out the music created using the ark. Certain elements that Kircher used notation to calculate (like determining vocal ranges by clef combinations and the size of the staff) we actually do in the Cogito module. This module is purely focused on output of complete music information to a music-notation language.

This module outputs to Lilypond, which could then be processed by that program to PDF, MIDI, or other formats.

This module also contains the central compose function that takes in a text processed by the Lectio module, and does the rest of the work of getting data from the ark, adjusting and translating that data into music (using Cogito), and then writing it.

Write individual data types to Lilypond strings

pitch2ly :: Pitch -> String

Write pitch as Lilypond music note. Look up needed string values for letter name, accidental, octave tick marks, and duration in lists based on data in given Pitch. If it is a Rest, just print the rest rhythm string.

Most of these just require using an enum value as index to a list of strings or characters. The octave requires us to calculate the number of commas or apostrophes to add (relative to Helmholtz octave 3 = c).

Write larger structures of music data to Lilypond, with necessary framing

enbrace

Arguments

:: String
string to put between enclosures

-> String
opening enclosure text (e.g., {)

-> String
closing text (e.g., })

-> String

Put things inside enclosing parens, brackets, etc.

lyMusicGroup :: String -> String

Group string in curly braces {string} with added newlines

lySimultaneousGroup :: String -> String

Group string in double angle brackets <<string>> with added newlines

voice2ly :: Voice -> ModeSystem -> LyricSection -> String

Write a Voice to a Lilypond music group:

\new Staff<< \new Voice { ... } >>

We have to include Lilypond midiInstrument here.

lyrics2ly :: LyricSection -> Voice -> String

Write a LyricSection to a Lilypond new Lyrics { } statement for a particular voice (VoiceName). Separate -- syllables with -- .

voice2lyOpening :: Voice -> String

The opening string per voice

voice2lyClosing :: Voice -> String

The end of the music part of the voice and beginning of the lyrics

voiceClosing :: Voice -> String

The end of the voice and its lyrics (discards input)

masterMusic2ly :: Arca -> [Symphonia] -> String

Convert all the music derived from input into notation output

makePreamble :: [String] -> String

Make Lilypond preamble of include commands

makeHeader :: ArkMetadata -> String

Make Lilypond header with the Arca as the author

All together now

compose :: Arca -> ArkMetadata -> [LyricSection] -> [SectionPerm] -> String

Set a prepared LyricSentence to music in one go. Return the text of a complete Lilypond file as a string. Write version number and preamble, and put music into score and StaffGroup (needed because we are doing Mensurstriche).