Cogito

Description

This module processes data from the ark to convert it into music (cogito, Latin, "I think").

Overview

This module receives input from the Lectio module in the form of a single list of one or more LyricSections, which contain the parsed text to be set to music and the parameters for setting it.

The module uses Kircher's rules to pull the appropriate data from the Arca musarithmica, that is, from the Arca built by the Aedifico module. It uses the Fortuna module to get lists of random permutation indices.

The main function is makeMusicScore, which applies all the necessary rules to select music data from the ark for each phrase of text, using the random permutations when a free choice would otherwise be needed. It takes the numerals and rhythmic symbols from Kircher's pinakes (rods); converts the numerals to pitches according to the mode, and combines the pitches and rhythms (accounting for rests as well).

The module creates the MusicScore data structure which contains all the data for the music in hierarchical sections that the Scribo module will convert to MEI XML.

Text underlay

Pitches and syllables are stored together in the Note type. In Syntagma I (simple syllabic counterpoint), we store one syllable for each note, skipping rests.

In Syntagma II, though, for florid counterpoint, Kircher does not specify how to underlay the text, and the settings have variable numbers of notes in the different voices and between different permutations in the same pinax. The underlay must be left to the human performer, then, and so we just lump all the lyrics for each phrase and put them under the first syllable as a textual incipit.

MEI vs. Lilypond output

We previously set up this module to feed data into the Scribo.Lilypond module, using the main function getSymphonia. It treated pitches and lyrics completely separately, which worked in Syntagma I but not in Syntagma II. These functions are still here at the bottom of the file. The test module still uses them, for now.

Pitches and Groups of Them

data RawPitch

A RawPitch is the same as a Pitch except instead of a Pnum it has an Int. This allows us to add and subtract from pitches and then use stdPitch to normalize them by adjusting the octave and pitch like the two digits of a base-7 number.

Constructors

RawPitch

Fields

rawPnum :: Int

rawOct :: Int

rawDur :: Dur

rawAccid :: Accid

rawAccidType :: AccidType

data Voice

A Voice is a list of pitches with an identifier for the voice type.

Constructors

Voice

Fields

voiceID :: VoiceName

Enum for Soprano, Alto, Tenor or Bass

music :: [Pitch]

mergeVoices :: [Voice] -> Voice

Make a list of voices with same voiceID into one Voice (combine the lists of Pitches into one)

type Chorus = [Voice]

A Chorus is a group (list) of four Voice items

TODO: We don't actually define it as being four items. TODO: Do we still need this with new MEI setup?

Manipulating the Pitch

newRest

Arguments

:: Dur
Rhythmic duration for this note

-> Pitch

Create a rest (that is, a Pitch with duration only)

We make a Pitch but set the pnum to Rest; oct and accid are set to special nil values (OctNil, AccidNil)

TODO: We are setting the octave using fromEnum OctNil: Isn't this the same as just setting it to zero? Is there a better way to mark this?

stdPitch :: RawPitch -> Pitch

Standardize pitch.

A Pitch is like a two-digit number in base 7, where the first digit (the "7s") is the octave. If the second digit (the "1s") is over 7, we must increment the octave. A "standardized pitch" in our implementation is one that conforms to this requirement, so the stdPitch method takes all the input for a Pitch and does the necessary conversions to set the octave.

We need this because Kircher's tables include both pitch number 1 and pitch 8. If pitch input is Kircher's pitch 8, then set pitch num to PcC (0) and add one to octave.

TODO: We don't check for values out of range because we know what input we are getting from the ark. Is this okay?

incPitch :: Pitch -> Int -> Pitch

Increment a pitch by increasing its Pnum pitch number and its octave if necessary (using Pitch structure like a base-7 number). Create a new Pitch with incremented Pnum and then standardize it with stdPitch to get correct octave and pitch number.

Adjust pitch for mode

modeMollis :: Mode -> ModeSystem -> Bool

Is a mode in cantus mollis? Should there be a flat in the key signature?

pnumAccidInMode :: Int -> Mode -> ModeList -> PnumAccid

Adjust a pitch to be in a given mode.

modalFinalInRange :: Mode -> ModeList -> VoiceName -> VoiceRanges -> Pitch

Get the modal final within range for this voice. What pitch = 0 in this mode? What is the lowest octave of that pitch permissible in the range for that voice? Set all the other pitches with reference to that, within the octave between 0--7

modalOctaveBase :: Mode -> ModeList -> VoiceName -> VoiceRanges -> Int

What octave is the modal final in for this voice's range?

accidentalShift :: Pitch -> Accid -> Pitch

Adjust the accidental either toward flats or toward sharps, within the Accid enum. If the accidental is unset we just return the original pitch.

flatten :: Pitch -> Pitch

Lower a pitch a semitone, but not past FlFl

sharpen :: Pitch -> Pitch

Raise a pitch a semitone, but not past ShSh

Measure distances between notes and correct bad intervals

Convert between diatonic and chromatic pitches to calculate intervals

absPitch :: Pitch -> Int

Convert Pitch to absolute pitch number

dia2chrom :: Pnum -> Int

Get chromatic offset from C for diatonic pitch classes

pEq :: Pitch -> Pitch -> Bool

Are two Pitches the same chromatic pitch, enharmonically equivalent?

pGt :: Pitch -> Pitch -> Bool

Pitch greater than?

pLt :: Pitch -> Pitch -> Bool

Pitch less than?

pGtEq :: Pitch -> Pitch -> Bool

Pitch greater than or equal?

pLtEq :: Pitch -> Pitch -> Bool

Pitch less than or equal?

absPitch7 :: Pitch -> Int

Absolute diatonic pitch (octave + pitch)

p7diff :: Pitch -> Pitch -> Int

Difference between pitches, diatonic interval

octaveAdjust :: Pitch -> (Int -> Int) -> Pitch

Copy a Pitch (unchanged if Rest), with given function applied to the octave member

octaveUp :: Pitch -> Pitch

Raise the octave by 1

octaveDown :: Pitch -> Pitch

Lower the octave by 1

Match pitches and rhythms

Get music data for a single voice

isRest :: Dur -> Bool

Check to see if a rhythmic duration is a rest type (the rest enums begin with LgR so we compare with that)

isPitchRest :: Pitch -> Bool

Is the Pitch a rest?

zipFill

Arguments

:: [a]
list 1

-> [b]
list 2

-> (a -> Bool)
test

-> b
substitute element to use instead of list 2 item, if list 1 item meets test

-> [(a, b)]
return list of pairs made of list 1 and 2, in item order

Take two lists and zip them together, that is, make a new list of ordered pairs constructed from sequential items of the two lists, BUT: if an element in the first list satisfies a given test, make a pair of that element and a sustitute element (sub) instead of the corresponding element from the second list.

pair2Pitch

Arguments

:: VoiceName

-> VoiceRanges

-> Mode

-> ModeList

-> ModeSystem

-> (Dur, Int)
duration and pitch number 0-7

-> Pitch

Make a pitch from duration and pitch number, getting octave based on voice name using voice2octave. If the duration input is a rest type (e.g., BrR), then make a newRest; otherwise a stdPitch.

Adjust the pitch for mode (and thereby standardize it) (pnumAccidInMode). Adjust the octave to put the pitch in the right range for the voice (adjustPitchInRange).

TODO: This could also be generalized; we are not checking inputs because we control data input.

isBflatInSignature :: Pnum -> Accid -> Mode -> ModeSystem -> Bool

Is this note a B flat, and if so, is the flat already in the key signature?

voice2octave :: Num p => VoiceName -> p

Get the right starting octave range for each voice type voice2octave :: VoiceName -> Int

pitchTooLow :: Pitch -> VoiceName -> VoiceRanges -> Bool

Is the pitch below the bottom limit of the voice range?

pitchTooHigh :: Pitch -> VoiceName -> VoiceRanges -> Bool

Is the pitch above the upper limit of the voice range?

lowRange :: VoiceName -> VoiceRanges -> Pitch

Get the bottom limit of the voice range

highRange :: VoiceName -> VoiceRanges -> Pitch

Get the top limit of the voice range

adjustPitchInRange :: Pitch -> VoiceName -> VoiceRanges -> Pitch

Adjust a pitch to be in the correct voice range (using Aedifico.vocalRanges). If it's in the right range for the voice, leave it alone; if it's too low raise it by an octave, or vice versa if it's too high; keep shifting octaves till it's in range.

isPitchInRange :: Pitch -> VoiceName -> VoiceRanges -> Bool

Is the Pitch within the proper range for its voice?

Adjust list of pitches to avoid bad intervals

stepwiseInRange :: [Pitch] -> VoiceName -> VoiceRanges -> [Pitch]

Go through list of pitches and reduce intervals that are more than a seventh

stepwiseVoiceInRange :: Voice -> VoiceRanges -> Voice

Adjust a whole Voice stepwise

stepwise :: [Pitch] -> [Pitch]

Go through list of pitches and reduce intervals that are more than a seventh

stepwiseVoice :: Voice -> Voice

Adjust a whole Voice stepwise

unleap :: Pitch -> Pitch -> Pitch

Reduce leap of more than a seventh by shifting octave of second note up or down until the interval is within range

TODO : - but what if after adjusting for leaps, a note is out of range? - and what if there is a descending scale that goes out of range and the only way to adjust it is to make a seventh? need to adjust a whole phrase

pitchMax :: [Pitch] -> Maybe Pitch

Return the highest pitch in a list of pitches.

pitchMin :: [Pitch] -> Maybe Pitch

Return the lowest pitch in a list of pitches.

voiceInRange :: Voice -> VoiceRanges -> Voice

TODO write your own max/min functions for pitches that ignore Rests

Adjust a whole Voice to be in range: check the highest and lowest notes in the list, compare to the range for the voice, and shift the whole thing by octave until all are in range; return error if it can't be done

All together: From input parameters to music

ark2voice

Arguments

:: Arca
ark data structure

-> ArkConfig
we pass this along to getVoice and getRperm; we use the Mode for pair2Pitch

-> PenultLength
penultimate syllable length

-> Int
syllable count

-> Int
line count

-> VoiceName
voice name enum

-> Perm
contains random index for voice and rhythm permutation

-> Voice

Central functions of the ark: given all parameters required by Kircher (style, meter, syllable count, penultimate syllable length), select a voice permutation (Kircher's number tables) from the appropriate part of the ark and match it to a rhythm permutation (his tables of note values).

Return a Voice with the pitches for a single voice part.

We use getVoice and getRperm from the Aedifico module.

Because the rhythms can include rest, we have to match up pitches and rhythms accordingly using zipFill with the test isRest.

Data structures to store music composed by the ark

data Note

A Note contains a pitch and a syllable, equivalent to MEI note

Constructors

Note

Fields

notePitch :: Pitch

noteSyllable :: Syllable

data Syllable

A Syllable is a single syllable to be paired with a Pitch, including its position in the word.

Constructors

Syllable

Fields

sylText :: String

sylPosition :: SyllablePosition

data SyllablePosition

What is the position of the syllable relative to the word? Beginning, middle, or end? This determines hyphenation.

Constructors

First

Middle

Last

Only

Tacet
no syllable

data MusicPhrase

A MusicPhrase contains all the notes set using one permutation drawn from the ark, for a single voice.

Constructors

MusicPhrase

Fields

phraseVoiceID :: VoiceName

notes :: [Note]

type MusicSentence = [MusicPhrase]

A list of MusicPhrase items

data MusicSection

A MusicSection contains all the music for one section in the input XML document, for a single voice, together with the parameters set in the input file.

Constructors

MusicSection

Fields

secVoiceID :: VoiceName

secConfig :: ArkConfig

secSentences :: [MusicSentence]

data MusicChorus

A MusicChorus is a four-voice SATB structure of MusicSection data. TODO do we really need it to be structured this way?

Constructors

MusicChorus

Fields

soprano :: MusicSection

alto :: MusicSection

tenor :: MusicSection

bass :: MusicSection

type MusicScore = [MusicChorus]

The full MusicScore is a list of SATB MusicChorus structures.

Methods to create and populate these data structures

makeSyllables :: Verbum -> [Syllable]

Take a Verbum read from the input file and turn it into a list of Syllables for storage in Notes. Record the syllable's position within the word.

blankSyllable :: Syllable

Just a blank syllable for filler when needed

makeMusicPhrase :: Arca -> ArkConfig -> VoiceName -> LyricPhrase -> Perm -> MusicPhrase

Compose the music for a whole LyricPhrase with one permutation from the ark, and package it into a MusicPhrase. Note that this is for a single voice only, not the four SATB voices. Line up pitches and syllables, skipping rests. In Syntagma I, line up text and notes syllabically (one syllable per note); in syntagma II (florid), lump the text into a single syllable and put it as an incipit text at the beginning of the phrase. (See module description for why Kircher's specification makes this is necessary.)

makeMusicSentence :: Arca -> ArkConfig -> VoiceName -> LyricSentence -> SentencePerm -> MusicSentence

Compose music for a LyricSentence for a single voice.

makeMusicChorus :: Arca -> LyricSection -> SectionPerm -> MusicChorus

Compose music for all four SATB voices for one LyricSection.

makeMusicSection :: Arca -> LyricSection -> SectionPerm -> VoiceName -> MusicSection

Put together all the music information for one LyricSection, for a single voice.

For a single voice:

extract ArkConfig for whole section

for each sentence in section:

extract list of perms, one per phrase

extract list of lyric phrases

apply same ArkConfig

for each phrase in sentence:

look up vperm according to config and perm

(for some pinakes, choose column by stanza = section num)

look up rperm according to config and perm

(for syntagma II, use same perm)

convert vperm nums to pitch names

(adjust pitches)

make Pitches: match pitches and rhythms, accounting for rests

match Notes: match each Pitch with PhraseVerbumSyllable according to syntagma

return a MusicPhrase

inside a MusicSentence

inside a MusicSection

makeMusicScore :: Arca -> [LyricSection] -> [SectionPerm] -> MusicScore

Compose the music for the whole document as a MusicScore, pulling the data from the Arca.

getChorus

Arguments

:: Arca
ark data structure

-> ArkConfig
we pass this along to ark2voice

-> LyricPhrase
input text processed by Lectio

-> Perm
Perm (from Fortuna, includes index for voice and rhythm)

-> Chorus

Get music data for all four voices and pack them into a Chorus.

TODO: Why do we use a Fortuna Perm here but just two integers in the preceding function? Do we need to have all these parameters for both of these functions?

We should be getting a new Perm for each Chorus.

setVoiceInitialRange :: Voice -> VoiceRanges -> Voice

Set the starting note of a voice to be within the proper range

data Symphonia

A Symphonia is the amalgamation of a list of Choruses into one Chorus

Constructors

Symphonia

Fields

chorus :: Chorus

lyricSection :: LyricSection

mergeChoruses :: [Chorus] -> Chorus

Turn [[S, A, T, B], [S1, A1, T1, B1]] into [[S, S1], [A, A1], [T, T1], [B, B1]]

getSymphonia :: Arca -> LyricSection -> SectionPerm -> Symphonia

To make a Symphonia we take a LyricSentence and list of Perms, use getChorus to get the ark data for each LyricPhrase in the sentence, each using its own Perm; then we use transpose to reorder the lists.

Where we had [[S1, A1, T1, B1], [S2, A2, T2, B2], [S3, A3, T3, B3]] we end with [[S1, S2, S3], [A1, A2, A3], [T1, T2, T3], [B1, B2, B3]]. A list of four voices becomes four lists of voices. We need to combine each of those voices into a single voice to have a list of four (longer) voices.

Adjust music of merged voices to avoid bad intervals (stepwise).

The Scribo module calls this function to get all the ark data needed to set a whole LyricSentence, in the central function of our implementation, Scribo.compose.

getMasterMusic :: Arca -> [LyricSection] -> [SectionPerm] -> [Symphonia]

Get all the music for the sections from input