Cogito.Ficta

Description

This module provides functions to adjust music created by the ark to follow conventions of musica ficta.

Adjust accidentals of individual pitches

changeNotePitch :: Note -> Pitch -> Note

Copy a Note but change the Pitch

adjustNotePitch :: (Pitch -> Pitch) -> Note -> Note

Copy a Note but adjust just its Pitch according to a function.

accidentalShift :: Pitch -> Accid -> Pitch

Adjust the accidental either toward flats or toward sharps, within the Accid enum. If the accidental is unset we just return the original pitch. We are treating all accidental shifts as musica ficta and giving a Suggested accidType.

flatten :: Pitch -> Pitch

Lower a pitch a semitone, but not past FlFl

sharpen :: Pitch -> Pitch

Raise a pitch a semitone, but not past ShSh

changeAccid :: Pitch -> Accid -> AccidType -> Pitch

Copy pitch but change accid and accidType

cancel :: Pitch -> Pitch

Cancel an accidental (suggested)

noteCancel :: Note -> Note

Cancel the Pitch within a Note.

fictaAccid :: Pitch -> Pitch

Make accidType Suggested

writeAccid :: Pitch -> Pitch

Make accidType Written

Utilities to adjust pitches by MusicSection

adjustNotesInSection :: ([Note] -> [Note]) -> MusicSection -> MusicSection

Do something to the Notes in every MusicPhrase within a MusicSection

changeNotesInPhrase :: MusicPhrase -> [Note] -> MusicPhrase

Copy a MusicPhrase but with new notes

adjustNotesInPhrase :: ([Note] -> [Note]) -> MusicPhrase -> MusicPhrase

Do something to the Note list in a MusicPhrase

fixFictaInSection

Arguments

:: ([Note] -> Note -> [Note])
fold function (arguments: stack and new item)

-> MusicSection

-> MusicSection

Fold a ficta-adjusting function over a MusicSection

fixFictaInPhrase

Arguments

:: ([Note] -> Note -> [Note])
fold function (arguments: stack and new item)

-> MusicPhrase

-> MusicPhrase

Fold a ficta-adjusting function over a MusicPhrase

foldStack

Arguments

:: ([a] -> a -> [a])
fold function (arguments: stack and new item)

-> [a] -> [a]

Generate a stack folding function to process a list

findCounterpoint

Arguments

:: MusicPhrase
phrase to search for counterpoint

-> MusicPhrase
phrase containing the point to match up in the other voice

-> Int
index of point in its phrase

-> Note

Compare two MusicPhrases and find the note in the one (lower) voice that coincides rhythmically with a given note in the other (upper) voice. Used to find harmonies and test them for bad intervals.

We find the top note by index and add the durations up to that point, then we add the durations in the bottom voice up to each item (that is, a scan) and then stop at the first item that matches the elapsed duration of the top voice.

durQuantity :: Dur -> Int

What is the elapsed time of a Dur in units where Fs (fusa) = 1?

Apply ficta adjustments to whole MusicPhrases

adjustFictaChorus :: ModeSystem -> ModeList -> MusicChorus -> MusicChorus

Adjust a chorus for musica ficta: adjust bass first, and then apply a function to each of the upper voices adjusting it relative to the revised bass.

adjustPhrasesInSection

Arguments

:: (MusicPhrase -> MusicPhrase -> MusicPhrase)
phrase transform function

-> MusicSection
lower voice section

-> MusicSection
upper voice section

-> MusicSection

Map a function to the phrases in one section (upper voice) relative to the phrases in another section (lower voice).

adjustFictaPhrase :: ModeList -> Mode -> MusicPhrase -> MusicPhrase -> MusicPhrase

Apply musica ficta adjustments to a whole MusicPhrase. Some adjustments only involve this phrase, while others must be relative to the bass voice. The bass voice should be adjusted already before this function is called.

Adjustments: (1) scale degree seven should be sharp if that is suggested in the mode table only if it is going up to eight and if the bass is on scale degree five; (2) scale degree six should be flat if that is suggested in the mode table only if it is going down to five; (3) repeated notes should match the last accidental in the series [as long as the harmonies with the bass are okay? TODO]; (4) there should be no cross-relations or augmented fifths.

TODO active development; not everything working

Specific adjustments by rule

scaleDegree :: ModeList -> Mode -> Pitch -> Int

Return the 0-indexed scale degree of a given pitch in a given mode (scale degree 0 is the modal final)

isFictaAccid :: Accid -> Pitch -> Bool

Does this Pitch have the given accidental as Suggested?

isFictaAccidNote :: Accid -> Note -> Bool

Apply isFictaAccid to a Note

fixIntervals

Arguments

:: Note
lower note

-> Note
upper note

-> Note
adjusted upper note

No cross relations, augmented fifths, or tritones between upper and lower note

fixUpperVoiceIntervals :: Note -> Note -> Note

Avoid cross relations (TODO other intervals?) between upper voices. If there is a cross relation on notes with Suggested accidentals, match the lower voice.

isCrossRelation :: Pitch -> Pitch -> Bool

Are these pitches the same pitch class but different accidentals? (E.g., F vs F#?)

isAugFifth

Arguments

:: Pitch
lower pitch

-> Pitch
higher pitch

-> Bool

Are these pitches an augmented fifth apart?

isTritone

Arguments

:: Pitch
lower pitch

-> Pitch
upper pitch

-> Bool

Are these two notes a tritone apart chromatically?

adjustBassFicta :: ModeSystem -> ModeList -> Mode -> MusicSection -> MusicSection

Adjust musica ficta accidentals in the bass.

fixMelodicTritone :: ModeSystem -> Mode -> [Note] -> Note -> [Note]

Fix melodic tritone before or after B (fold function)

fixSixSeven :: ModeList -> Mode -> [Note] -> Note -> [Note]

Fix descending scale-degree sevens and ascending sixes (fold function)

cancelNote :: Note -> Note

Cancel the Pitch in a Note

fixFlatSharp :: [Note] -> Note -> [Note]

If this note is suggested flat, and the next note is suggested sharp, make this note natural.

fixAccidRepeat :: [Note] -> Note -> [Note]

If there are two subsequent accidental inflections of a note (e.g., F natural--F sharp or vice versa), make the first note match the second. This deals with a byproduct of sharpLeadingTone with repeated notes, where that function would turn F--F--G into F--F#--G. This function would make it F--G.

fixLeadingTonesInPhrase

Arguments

:: ModeList

-> Mode

-> MusicPhrase
bass voice

-> MusicPhrase
upper voice to be adjusted

-> MusicPhrase
adjusted upper voice

Fix the leading tones of a MusicPhrase.

fixLeadingTonePairs

Arguments

:: ModeList

-> Mode

-> [(Note, Note)]
stack of pairs of upper and lower notes

-> (Note, Note)
next upper voice note pair

-> [(Note, Note)]
pairs with adjusted upper notes

Fold function for adjusting the leading tones in the upper voice of a list of (upper, lower) Note pairs

sharpLeadingTone

Arguments

:: ModeList
list (table) of modes from the ark

-> Mode
current mode

-> Note
bass note

-> Note
current note to be adjusted

-> Note
next note

-> Note

Sharp highest scale degree in an upper voice (1) when it leads up to scale-degree eight, and (2) when the bass note is on scale degree five. (TODO or two?)

Remember, these are all 0-indexed numbers instead of the 1-indexed numbers used in speech (degree 6 here is "scale degree 7" in speech).